refactor: test code에서 동작할 mock S3Service 리팩토링
